Split app.js phase 2b: extract settings and apps-picker modules
In-app tests pass (132/132).
